Upgraded parent pom
