+++ pom.xml
@@ -22,7 +22,7 @@
     <parent>
         <artifactId>dans-dropwizard-project</artifactId>
         <groupId>nl.knaw.dans.shared</groupId>
-        <version>7.2.0</version>
+        <version>7.4.0-SNAPSHOT</version>
     </parent>
 
     <groupId>nl.knaw.dans</groupId>
@@ -52,17 +52,14 @@
         <dependency>
             <groupId>nl.knaw.dans</groupId>
             <artifactId>dans-java-utils</artifactId>
-            <version>0.0.1</version>
         </dependency>
         <dependency>
             <groupId>nl.knaw.dans</groupId>
             <artifactId>dans-dataverse-client-lib</artifactId>
-            <version>0.1.0</version>
         </dependency>
         <dependency>
-            <groupId>io.dropwizard.modules</groupId>
+            <groupId>io.dropwizard</groupId>
             <artifactId>dropwizard-health</artifactId>
-            <version>1.7.2</version>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
