+++ src/main/java/nl/knaw/dans/virusscan/health/ClamdHealthCheck.java
@@ -13,37 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.virusscan.core.health;
+package nl.knaw.dans.virusscan.health;
 
 import com.codahale.metrics.health.HealthCheck;
-import nl.knaw.dans.virusscan.core.service.DataverseApiService;
+import nl.knaw.dans.virusscan.core.service.ClamdService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
-public class DataverseHealthCheck extends HealthCheck {
-    private static final Logger log = LoggerFactory.getLogger(DataverseHealthCheck.class);
+public class ClamdHealthCheck extends HealthCheck {
+    private static final Logger log = LoggerFactory.getLogger(ClamdHealthCheck.class);
 
-    private final DataverseApiService dataverseApiService;
+    private final ClamdService clamdService;
 
-    public DataverseHealthCheck(DataverseApiService dataverseApiService) {
-        this.dataverseApiService = dataverseApiService;
+    public ClamdHealthCheck(ClamdService clamdService) {
+        this.clamdService = clamdService;
     }
 
     @Override
     protected Result check() {
         try {
-            var info = dataverseApiService.getDataverseInfo();
+            var result = clamdService.ping();
+            log.trace("Result from ClamAV PING request: {}", result);
 
-            if (info.getStatus().equalsIgnoreCase("OK")) {
+            if ("PONG\n".equalsIgnoreCase(result)) {
                 return Result.healthy();
             }
             else {
-                throw new IOException(String.format("Version request returned incorrect status '%s'", info.getStatus()));
+                throw new IOException(String.format("Unexpected output from ClamAV: %s", result));
             }
         }
         catch (IOException e) {
+            log.error("IO error occurred while communicating with ClamAV", e);
             return Result.builder()
                 .withMessage(e.getMessage())
                 .unhealthy(e)
