+++ src/main/java/nl/knaw/dans/virusscan/health/DataverseHealthCheck.java
@@ -13,39 +13,37 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.virusscan.core.health;
+package nl.knaw.dans.virusscan.health;
 
 import com.codahale.metrics.health.HealthCheck;
-import nl.knaw.dans.virusscan.core.service.ClamdService;
+import nl.knaw.dans.virusscan.core.service.DataverseApiService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
-public class ClamdHealthCheck extends HealthCheck {
-    private static final Logger log = LoggerFactory.getLogger(ClamdHealthCheck.class);
+public class DataverseHealthCheck extends HealthCheck {
+    private static final Logger log = LoggerFactory.getLogger(DataverseHealthCheck.class);
 
-    private final ClamdService clamdService;
+    private final DataverseApiService dataverseApiService;
 
-    public ClamdHealthCheck(ClamdService clamdService) {
-        this.clamdService = clamdService;
+    public DataverseHealthCheck(DataverseApiService dataverseApiService) {
+        this.dataverseApiService = dataverseApiService;
     }
 
     @Override
     protected Result check() {
         try {
-            var result = clamdService.ping();
-            log.trace("Result from ClamAV PING request: {}", result);
+            var info = dataverseApiService.getDataverseInfo();
 
-            if ("PONG\n".equalsIgnoreCase(result)) {
+            if (info.getStatus().equalsIgnoreCase("OK")) {
                 return Result.healthy();
             }
             else {
-                throw new IOException(String.format("Unexpected output from ClamAV: %s", result));
+                throw new IOException(String.format("Version request returned incorrect status '%s'", info.getStatus()));
             }
         }
         catch (IOException e) {
-            log.error("IO error occurred while communicating with ClamAV", e);
             return Result.builder()
                 .withMessage(e.getMessage())
                 .unhealthy(e)
