+++ src/test/java/nl/knaw/dans/virusscan/health/ClamdHealthCheckTest.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.virusscan.core.health;
+package nl.knaw.dans.virusscan.health;
 
 import nl.knaw.dans.virusscan.core.service.ClamdService;
+import nl.knaw.dans.virusscan.health.ClamdHealthCheck;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
